requirements.txt
fire
jax
jaxlib
librosa
matplotlib
numpy
opax
pax3
pooch
pytest
tensorflow
tqdm